Add nested slogan class to banner
There is a slogan on the page, which is a child of the banner class.
The banner class is already represented as a class in the SCSS file,
but slogan was not. Since slogan is a child of banner, its specific
styling was nested inside of banner to show clear DOM relationships
within the styling sheet.