Add custom color variable
A custom color for the HTML page will be used repeatedly, which is
tedious an unneccesary to continuously define and easy to mess up.
Therefore this custom color, translucent-white, was saved as a
variable at the top of the stylesheet. The color can be called
anywhere within the stylesheet using the variable's name and if the
color needed to be changed, you only need to update the value in the
definition at the top of the stylesheet instead of at each instance
where the color is used.