Add standard border value
The same list is used multiple times throughout the SS as the value
of a border property, which means that should be converted to a
variable so that the variable becomes an easily reusable piece of
code that can be changed very simply. The list of 4px, solid, black
was set as the value for the standard-border variable.